feat: provide guidance in browser console when logging `$state` objects (#13142)
* feat: provide guidance in browser console when logging `$state` objects

Wrap console.log/warn/error statements in DEV mode with a check whether or not they contain state objects. Closes #13123

This is an alternative or enhancement to #13070. Alternative if we deem it the better solution. Enhancement because it's not as robust as a custom formatter: We only check the top level of each entry (though we could maybe traverse a few levels), and if you're logging class instances, snapshot currently stops at the boundaries there and so you don't get snapshotted values for these (arguably this is a more general problem of $inspect and $state.snapshot), whereas with custom formatter it doesn't matter at which level you come across it.

* lint

* use normal warning mechanism, so we can link to docs etc

* add a few more methods

---------

Co-authored-by: Rich Harris <rich.harris@vercel.com>