Correct create svelte 4 project command (#14498)
People visiting v4.svelte.dev are interested in pre-5 svelte.
Latest version of `create-svelte` package provides no functionality to create a project (it's a notice to use v5). 
Provided materials created for v4 and earlier need not be simply thrown away, creating a pre-v5 project should still be possible.
`6.4.0` seems to be the last version of `create-svelte` package that can be used to create a v4 svelte project.