fix: fire delegated events on target even it was disabled in the meantime
They DOM could've been updated already by the time the event bubbling logic is reached, so the target element itself could not be notified of the event because it is marked as disabled. But the target could not have been disabled because it emits the event in the first place, so we emit regardless.

No test because not reproducible in jsdom environment.

Fixes #15256