docs: Improve example code in - When not to use $effect - section
The main change treats spent as the source of truth and derives left from it—simplifying the code, reducing the risk of inconsistencies, and improving the clarity of their relationship.

I also changed let total = 100 to const total = 100 to more clearly convey that the value is intended to remain constant.