docs: Improve sample code in `When not to use $effect` section (#15994)
* docs: Improve example code in - When not to use $effect - section

The main change treats spent as the source of truth and derives left from it—simplifying the code, reducing the risk of inconsistencies, and improving the clarity of their relationship.

I also changed let total = 100 to const total = 100 to more clearly convey that the value is intended to remain constant.

* use normal binding for spent

* change let to const in prior example

* use diff highlighting

* i accidentally a typo

---------

Co-authored-by: Rich Harris <rich.harris@vercel.com>