feat: add parent hierarchy to `__svelte_meta` objects at dev time (#16255)
* feat: add parent hierarchy to `__svelte_meta` objects at dev time

This adds a `parent` property to the `__svelte_meta` properties that are added to elements at dev time. This property represents the closest non-element parent the element is related to. For example for `{#if ...}<div>foo</div>{/if}` the `parent` of the div would be the line/column of the if block.
The parent is recursive and goes upwards (through component boundaries) until the root component is reached, which has no parent.

part of #11389

* oops

* Apply suggestions from code review

Co-authored-by: Rich Harris <rich.harris@vercel.com>

* tweak

* original component tag

* make render appear in tree, keep tree in sync when rerenders occur

---------

Co-authored-by: Rich Harris <rich.harris@vercel.com>