fix: traverse graph when time traveling
Often deriveds are currently unowned but through graph traversal during `is_dirty()` might reconnect to the graph, therefore having dependent effects register themselves as reactions.

The problem prior to this fix was that the graph was not traversed to update the derived's dependencies and unowned status of deriveds in batch_values - the batch value was just returned -, which results in reactivity loss. https://github.com/sveltejs/svelte/issues/17024#issuecomment-3476761662 contains a detailed rundown of such a case. The fix is to still traverse the graph, though not executing any deriveds in the process.

Fixes #17024
Fixes https://github.com/sveltejs/svelte/issues/17049#issuecomment-3459910165 (and therefore everything that was still buggy in that issue I think)