chore: remove `UNOWNED` flag (#17105)
Fixes #17024
Fixes #17049 (comment) (and therefore everything that was still buggy in that issue I think)

* chore: remove unowned check when calling `e.effect_in_unowned_derived`

* WIP

* all non-unit tests passing

* tidy

* WIP

* WIP

* WIP

* note to self

* fix

* fix

* hmm maybe not

* try this

* simplify

* remove skip_reaction

* docs

* add changeset, in case this results in changed behaviour

* Update packages/svelte/src/internal/client/reactivity/effects.js

Co-authored-by: Simon H <5968653+dummdidumm@users.noreply.github.com>

* fix #17024

* fix comment

* revert

* fix

* dry

* changeset

* fix WAS_MARKED logic

* failing test (that uncovered other unrelated bug) + fix

* fix: delete from batch_values on updates (#17115)

* fix: delete from batch_values on updates

This fixes a bug where a derived would still show its old value even after it was indirectly updated again within the same batch. This can for example happen by reading a derived on an effect, then writing to a source in that same effect that makes the derived update, and then read the derived value in a sibling effect - it still shows the old value without the fix.

The fix is to _delete_ the value from batch_values, as it's now the newest value across all batches. In order to not prevent breakage on other batches we have to leave the status of deriveds as-is, i.e. within is_dirty and update_derived we cannot update its status. That might be a bit more inefficient as you now have to traverse the graph for each `get` of that derived (it's a bit like they are all disconnected) but we can always optimize that later if need be.

Another advantage of this fix is that we can get rid of duplicate logic we had to add about unmarking and reconnecting deriveds, because that logic was only needed for the case where deriveds are read after they are updated, which now no longer hits that if-branch

* keep derived cache, but clear it in mark_reactions (#17116)

---------

Co-authored-by: Rich Harris <rich.harris@vercel.com>
Co-authored-by: Simon H <5968653+dummdidumm@users.noreply.github.com>
Co-authored-by: Simon Holthausen <simon.holthausen@vercel.com>