fix: keep deriveds reactive after their original parent effect was destroyed (#17171)
Use case: Remote queries that are created on one screen, then are used again on another screen. Original parent effect is destroyed in that case but derived should still be reactive. It wasn't prior to this fix because inside `get` the `destroyed` variable would be true and so deps would not properly be recorded.

Fixes https://github.com/sveltejs/kit/issues/14814