fix: don't execute attachments and attribute effects eagerly (#17208)
* fix: don't execute attachments and attribute effects eagerly

attributes_effect and attachments are blocks since they need the managed "don't just destroy children effects"-behavior, but they're not block effects in the sense of "run them eagerly while traversing the effect tree or while flushing effects". Since the latter was the case until now, it meant that forks could cause visible UI updates.

This PR introduces a new flag to fix that. `BLOCK_NON_EAGER` is basically a combination of block effects (with respects to the managed behavior) and render effects (with respects to the execution timing).

Fixes https://github.com/sveltejs/kit/issues/14931

* managed_effect