fix: link offscreen items and last effect in each block correctly (#17240)
* fix: link offscreen items and last effect in each block correctly

It's possible that due to how new elements are inserted into the array that `effect.last` is wrong. We need to ensure it is really the last item to keep items properly connected to the graph. In addition we link offscreen items after all onscreen items, to ensure they don't have wrong pointers.

Fixes #17201

* revert #17244

* add test